fix: remove Bitcoin-specific witness code and improve logging
- Remove inappropriate wtxid logging throughout codebase
- Simplify orphan transaction processing by removing witness-specific logic
- Fix txorphanage logging to show correct orphan child hash
- Update uint64_t casting for consistency in logging.h
- Modernize Python test strings to use f-strings
- Update test expectations to match simplified logging

Addresses CodeRabbit feedback while preserving Bitcoin intent for Dash

+++ src/logging.h
@@ -67,8 +67,8 @@ namespace BCLog {
 #endif
         BLOCKSTORE  = (1 << 26),
         TXRECONCILIATION = (1 << 27),
-        SCAN        = (1 << 28),
-        TXPACKAGES  = (1 << 29),
+        SCAN        = (uint64_t)1 << 28,
+        TXPACKAGES  = (uint64_t)1 << 29,
 
         //Start Dash
         CHAINLOCKS  = ((uint64_t)1 << 32),
